requirements.txt
asgiref==3.5.2
attrs==22.1.0
Automat==20.2.0
backports.zoneinfo==0.2.1
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
confusable-homoglyphs==3.2.0
Django==4.0.6
django-cors-headers==3.13.0
django-registration==3.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
mypy-extensions==0.4.3
parsel==1.6.0
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
psycopg2-binary==2.9.3
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.2
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tldextract==3.3.1
tomli==2.0.1
Twisted==22.4.0
typing_extensions==4.3.0
urllib3==1.26.11
w3lib==1.22.0
zope.interface==5.4.0
